requirements.txt
google-generativeai
google-cloud-texttospeech
flask
flask-cors
python-dotenv
numpy
gunicorn